Convert expression to a package and split them into simple and compound modules
